+++ src/plugins/gta3/std.data/main.cpp
@@ -7,6 +7,17 @@
 #include "data.hpp"
 using namespace modloader;
 
+/*
+ *  !!!! data_traits important implementation information !!!!
+ *
+ *   [*] It's better that each data trait (i.e. for each file type) be contained in it's own cpp, see next * for a reason
+ *   [*] Note data.hpp / cache.hpp should not be included in a PCH but in the data traits cpp, that because the magic for each cache
+ *       is the hash of the compilation time of the trait translation unit, that means it is a static (by unit) function in cache.hpp
+ *       meaning including it by PCH (stdinc.hpp) would defeat it purposes of one hash for each translation unit for each time it compiles.
+ *       
+ */
+
+
 
 DataPlugin plugin;
 REGISTER_ML_PLUGIN(::plugin);